refactor the ApptCommandParsers into a new package
