+++ src/main/java/seedu/address/logic/parser/AppointmentCommandParser/ShowApptCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.AppointmentCommandParser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
